requirements.txt
# Core ML Framework
torch>=2.5.0
torchvision>=0.20.0
pytorch-lightning>=2.4.0
transformers>=4.45.0
accelerate>=0.34.0
peft>=0.13.0
torchmetrics>=1.4.0

# 3D Medical Imaging
monai>=1.4.0
torchio>=0.19.0
nibabel>=5.2.0

# Video & Vision
timm>=1.0.9
einops>=0.8.0

# Scientific Computing
numpy>=2.0.0
scipy>=1.13.0
scikit-image>=0.24.0
opencv-python>=4.10.0
tifffile>=2024.8.30
pandas>=2.2.0

# Visualization
matplotlib>=3.9.0
seaborn>=0.13.2
plotly>=5.23.0

# Configuration & Experiment Tracking
hydra-core>=1.3.2
omegaconf>=2.3.0
wandb>=0.17.7

# Development & Testing
pytest>=8.3.2
pytest-cov>=5.0.0
black>=24.8.0
flake8>=7.1.0
mypy>=1.11.0

# Utilities
tqdm>=4.66.5
pyyaml>=6.0.2
pillow>=10.4.0
